ota-ng/swpkg/gen_swp_fb/requirements.txt
asn1==2.4.1
tinyec==0.3.1
python-magic==0.4.24
